Move sparse/swarmpack to unique dirs
