+++ src/CMakeLists.txt
@@ -217,21 +217,21 @@ add_library(parthenon
   outputs/restart_hdf5.hpp
 
   pack/block_selector.hpp
-  pack/make_pack_descriptor.hpp
-  pack/make_pack_descriptor.cpp
-  pack/make_swarm_pack_descriptor.hpp
-  pack/pack_descriptor.cpp
-  pack/pack_descriptor.hpp
   pack/pack_utils.hpp
-  pack/sparse_pack.hpp
-  pack/sparse_pack.cpp
-  pack/sparse_pack_base.cpp
-  pack/sparse_pack_base.hpp
-  pack/sparse_pack_cache.hpp
-  pack/sparse_pack_cache.cpp
-  pack/swarm_default_names.hpp
-  pack/swarm_pack.hpp
-  pack/swarm_pack_base.hpp
+  pack/default_names.hpp
+  pack/sparse_pack/make_pack_descriptor.hpp
+  pack/sparse_pack/make_pack_descriptor.cpp
+  pack/sparse_pack/pack_descriptor.cpp
+  pack/sparse_pack/pack_descriptor.hpp
+  pack/sparse_pack/sparse_pack.hpp
+  pack/sparse_pack/sparse_pack.cpp
+  pack/sparse_pack/sparse_pack_base.cpp
+  pack/sparse_pack/sparse_pack_base.hpp
+  pack/sparse_pack/sparse_pack_cache.hpp
+  pack/sparse_pack/sparse_pack_cache.cpp
+  pack/swarm_pack/make_swarm_pack_descriptor.hpp
+  pack/swarm_pack/swarm_pack.hpp
+  pack/swarm_pack/swarm_pack_base.hpp
   
   parthenon/driver.hpp
   parthenon/package.hpp
