+++ src/bvals/boundary_conditions_generic.hpp
@@ -28,9 +28,9 @@
 #include "mesh/domain.hpp"
 #include "mesh/mesh.hpp"
 #include "mesh/meshblock.hpp"
-#include "pack/make_pack_descriptor.hpp"
-#include "pack/sparse_pack.hpp"
-#include "pack/swarm_default_names.hpp"
+#include "pack/default_names.hpp"
+#include "pack/sparse_pack/make_pack_descriptor.hpp"
+#include "pack/sparse_pack/sparse_pack.hpp"
 
 namespace parthenon {
 namespace BoundaryFunction {
