+++ src/interface/meshblock_data.hpp
@@ -29,9 +29,9 @@
 #include "interface/variable.hpp"
 #include "interface/variable_pack.hpp"
 #include "mesh/domain.hpp"
-#include "pack/sparse_pack_base.hpp"
-#include "pack/sparse_pack_cache.hpp"
-#include "pack/swarm_pack_base.hpp"
+#include "pack/sparse_pack/sparse_pack_base.hpp"
+#include "pack/sparse_pack/sparse_pack_cache.hpp"
+#include "pack/swarm_pack/swarm_pack_base.hpp"
 #include "utils/concepts_lite.hpp"
 #include "utils/error_checking.hpp"
 #include "utils/unique_id.hpp"
