+++ src/interface/swarm_comms.cpp
@@ -19,7 +19,7 @@
 #include <vector>
 
 #include "mesh/mesh.hpp"
-#include "pack/swarm_default_names.hpp"
+#include "pack/default_names.hpp"
 #include "swarm.hpp"
 #include "utils/error_checking.hpp"
 #include "utils/sort.hpp"
