+++ src/pack/block_selector.hpp
@@ -31,7 +31,7 @@
 #include "interface/mesh_data.hpp"
 #include "interface/variable.hpp"
 #include "pack/pack_utils.hpp"
-#include "pack/sparse_pack_base.hpp"
+#include "pack/sparse_pack/sparse_pack_base.hpp"
 #include "utils/concepts_lite.hpp"
 #include "utils/type_list.hpp"
 #include "utils/utils.hpp"
