+++ src/pack/sparse_pack/sparse_pack.hpp
@@ -10,8 +10,8 @@
 // license in this material to reproduce, prepare derivative works, distribute copies to
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
-#ifndef PACK_SPARSE_PACK_HPP_
-#define PACK_SPARSE_PACK_HPP_
+#ifndef PACK_SPARSE_PACK_SPARSE_PACK_HPP_
+#define PACK_SPARSE_PACK_SPARSE_PACK_HPP_
 
 #include <algorithm>
 #include <utility>
@@ -19,7 +19,7 @@
 
 #include "coordinates/coordinates.hpp"
 #include "pack/block_selector.hpp"
-#include "pack/sparse_pack_base.hpp"
+#include "pack/sparse_pack/sparse_pack_base.hpp"
 #include "utils/concepts_lite.hpp"
 #include "utils/type_list.hpp"
 
@@ -444,4 +444,4 @@ inline SparsePack<Ts...> SparsePack<Ts...>::Descriptor::GetPack(
 
 } // namespace parthenon
 
-#endif // PACK_SPARSE_PACK_HPP_
+#endif // PACK_SPARSE_PACK_SPARSE_PACK_HPP_
