+++ src/pack/sparse_pack/sparse_pack_cache.cpp
@@ -15,7 +15,7 @@
 
 #include "interface/mesh_data.hpp"
 #include "interface/meshblock_data.hpp"
-#include "pack/sparse_pack_cache.hpp"
+#include "pack/sparse_pack/sparse_pack_cache.hpp"
 
 namespace parthenon {
