+++ src/pack/sparse_pack/sparse_pack_cache.hpp
@@ -10,8 +10,8 @@
 // license in this material to reproduce, prepare derivative works, distribute copies to
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
-#ifndef PACK_SPARSE_PACK_CACHE_HPP_
-#define PACK_SPARSE_PACK_CACHE_HPP_
+#ifndef PACK_SPARSE_PACK_SPARSE_PACK_CACHE_HPP_
+#define PACK_SPARSE_PACK_SPARSE_PACK_CACHE_HPP_
 
 #include <map>
 #include <string>
@@ -19,8 +19,8 @@
 #include <unordered_map>
 #include <vector>
 
-#include "pack/pack_descriptor.hpp"
-#include "pack/sparse_pack_base.hpp"
+#include "pack/sparse_pack/pack_descriptor.hpp"
+#include "pack/sparse_pack/sparse_pack_base.hpp"
 
 namespace parthenon {
 // Object for cacheing sparse packs in MeshData and MeshBlockData objects. This
@@ -61,4 +61,4 @@ class SparsePackCache {
 
 } // namespace parthenon
 
-#endif // PACK_SPARSE_PACK_CACHE_HPP_
+#endif // PACK_SPARSE_PACK_SPARSE_PACK_CACHE_HPP_
