+++ src/parthenon/package.hpp
@@ -27,12 +27,12 @@
 #include <mesh/mesh.hpp>
 #include <mesh/meshblock.hpp>
 #include <mesh/meshblock_pack.hpp>
-#include <pack/make_pack_descriptor.hpp>
-#include <pack/make_swarm_pack_descriptor.hpp>
-#include <pack/pack_descriptor.hpp>
 #include <pack/scratch_variables.hpp>
-#include <pack/sparse_pack.hpp>
-#include <pack/swarm_pack.hpp>
+#include <pack/sparse_pack/make_pack_descriptor.hpp>
+#include <pack/sparse_pack/pack_descriptor.hpp>
+#include <pack/sparse_pack/sparse_pack.hpp>
+#include <pack/swarm_pack/make_swarm_pack_descriptor.hpp>
+#include <pack/swarm_pack/swarm_pack.hpp>
 #include <parameter_input.hpp>
 #include <parthenon_manager.hpp>
 #include <utils/index_split.hpp>
